+++ src/git/queries/refs.rs
@@ -95,7 +95,7 @@ fn get_remote_name(parts: &[&str]) -> Option<String> {
   }
 }
 
-pub fn make_ref_info(info: RefInfoPart, commit_id: String, time: f32) -> RefInfo {
+pub fn make_ref_info(info: RefInfoPart, commit_id: String, time: i64) -> RefInfo {
   let RefInfoPart {
     id,
     location,
